pyproject.toml
[tool.poetry]
name = "brain-stroke-segmentation"
version = "0.1.0"
description = "Brain stroke segmentation using U-Net with EfficientNet-B4 encoder"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "brain_stroke_segmentation"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
torchvision = "^0.15.0"
pytorch-lightning = "^2.0.0"
segmentation-models-pytorch = "^0.3.3"
albumentations = "^1.3.0"
opencv-python = "^4.8.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
tqdm = "^4.65.0"
hydra-core = "^1.3.0"
omegaconf = "^2.3.0"
mlflow = "^2.8.0"
dvc = {extras = ["s3", "gs"], version = "^3.0.0"}
onnx = "^1.14.0"
onnxruntime = "^1.16.0"
onnxruntime-gpu = "^1.16.0"
fire = "^0.5.0"
gitpython = "^3.1.0"
pillow = "^10.0.0"
kagglehub = {version = "^0.2.0", optional = true}

[tool.poetry.extras]
kaggle = ["kagglehub"]

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.4.0"
prettier = "^3.0.0"

[tool.poetry.scripts]
train = "brain_stroke_segmentation.commands:train"
infer = "brain_stroke_segmentation.commands:infer"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv", "venv"]

